pyproject.toml
[tool.poetry]
name = "zero-trust-ai-defense"
version = "0.1.0"
description = "Zero-trust security agent model for detecting and countering AI/ML attacks"
authors = ["Security Team"]
readme = "README.md"
packages = [{include = "cake"}, {include = "main.py"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
pandas = "^2.0.0"
faker = "^19.0.0"
jinja2 = "^3.1.0"
aiohttp = "^3.9.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
redis = "^5.0.0"
cryptography = "^41.0.0"
python-dateutil = "^2.8.0"
requests = "^2.31.0"
stripe = "^6.0.0"
pyotp = "^2.9.0"
qrcode = "^7.4.2"
pillow = "^11.0.0"
anthropic = "^0.39.0"
rich = "^13.7.0"
docker = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ztai-defense = "main:main"
defense-agent = "agents.agent_cli:main"
